requirements.txt
Django==2.2.2
djangorestframework==3.9.4
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
gunicorn==19.6.0
